unifying resolvers and easier backends
