+++ gundb/backends/__init__.py
@@ -3,3 +3,4 @@ from .memory import Memory
 from .udb import UDB
 from .pickle import Pickle
 from .rediskv import RedisKV
+from .mongo import Mongo
